Changed ID field and got some requests working.

+++ src/main/java/edu/rosehulman/covidtracer/repository/PositivePoolRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import edu.rosehulman.covidtracer.model.PositivePool;
 
 @Repository
-public interface PositivePoolRepository extends JpaRepository<PositivePool, Integer> {
+public interface PositivePoolRepository extends JpaRepository<PositivePool, Long> {
 
 }
